requirements.txt
click
opencv-python
pyyaml
requests 